requirements.txt
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0